requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
anthropic==0.43.0
pydantic==2.10.4
requests==2.32.3
SQLAlchemy==2.0.37
psycopg2-binary==2.9.10
alembic==1.14.0